+++ packages/user-ui/src/pages/main/components/mod.rs
@@ -1,0 +1,2 @@
+pub mod institution_box;
+pub mod project_box;
