+++ packages/user-ui/src/pages/main/mod.rs
@@ -3,3 +3,5 @@ mod i18n;
 
 pub mod page;
 pub use page::MainPage;
+
+pub mod components;
